Added more attributes to transcription_index.json
After a lot of trial and error, that the easiest way to go about linking
the search results to video was to have it all in one json object.
Instead of multiple calls to multiple json objects. If this isn't the
cleanest way then will refactor later. Will focus on functionality for
now.