Add the `/home/user/` files back into filesystem
I used the `utils/buildfs` script to build a filesystem with the user
files in it, then moved that built fs into `sysroot/`. However,
because the build script and packages in `utils` are old, some of the
changes in the latest filesystem were overwritten by the old
filesystem. I kept only the `home/user/` file addition changes, and
discarded the rest. The `-p` flag to `git add` command was very
helpful.