Use full URLs as the filesystem base paths
Jor1k loads the *fs.json files to load metadata about the files
present in the filesystem, and then fetches the individual files based
on that metadata. The URLs to fetch files from are generated using the
filesystem "src" and the files' relative paths. No additional base URL
or path is added.

The "../sys/" fs "src" paths work for the Jor1k project because the
XHR requests for the files are made by the worker at
https://s-macke.github.io/jor1k/demos/jor1k-worker-min.js, and so
relative URLs are resolved relative to
https://s-macke.github.io/jor1k/demos/. For example, the URL for the
file "usr/lib/libc.so.bz2" is generated as
"../sys/fs/usr/lib/libc.so.bz2", which when fetched by the browser,
results in a request to
https://s-macke.github.io/jor1k/sys/fs/usr/lib/libc.so.bz2. Now, the
Jor1k source contains a submodule called "sys" at its root, which
points to the s-macke/jor1k-sysroot repository. GitHub Pages
automatically pulls in the submodule files when the Page is built.
This means that files available in the s-macke/jor1k-sysroot
repository will be available on the web through the
https://s-macke.github.io/jor1k/sys/ URLs. Which means that
https://s-macke.github.io/jor1k/sys/fs/usr/lib/libc.so.bz2 will point
to this file:
github.com/s-macke/jor1k-sysroot/blob/master/fs/usr/lib/libc.so.bz2.

I do not want to use a submodule yet for fetching the filesystem in
sysbuild, because it'll require cloning another repository during
development. I think it is easier to just point to a sysassets URL,
which means that file lookups should not be dependent on the "../sys/"
path resolving correctly, and so this commit modifies the fs "src"
paths to use the full URLs to sysassets.

At some point though, using a submodule will be a good idea, as it
will allow pointing to a particular commit in the filesystem
repository (sysassets). This will allow us to make changes to
sysassets without fear of breaking the live site, and is generally
more robust.